chore: Consolidating CI integration jobs to get CI to run faster
