+++ internal/azurehelper/storage_account_test.go
@@ -1,5 +1,3 @@
-//go:build azure
-
 package azurehelper_test
 
 import (
